web.ssl: Replace the ridiculously long cipher list with cipher classes.
 1. We don't need to exclude, include or enumerate DSS, ECDSA or RSA backed
    cipher suites, because if they are used or not is already determined by
    the user's certificate type.
 2. No need to name every cipher. OpenSSL nows sorted classes since 1.0.1,
    which Discourse's docker container already ships with.
 3. Using classes will enable usage of ChaCha20Poly1305, RC6, HC-128, Rabbit
    and other ciphers, if and where available.
 4. Class HIGH already excludes EXPORT and MD5.

ECDH+RC4+SHA is added because it had been enumerated in the old configuration
file. If you can, remove it in future versions.

ECDH includes ECDHE and ECDH. Therefore we need to remove ECDH retaining ECDHE
using !kECDH.

AES128 is preferred over AES256, hence the -AES256 and the repetition. Ciphers
won't be added twice, so we pull in combinations with AES256 by that.

!PSK because there's still variants which, even if no PSK is set, negotiate
with an empty one.