Revert "templates/web.sssl.template.yml: Disable spdy header compression"
This reverts commit c0ed494c3acf1acf629379f33f413d7660183855. It was late and I
misread the info here:

	https://github.com/18F/tls-standards/issues/24

The first comment clearly states that the server *can* compress spdy header, as
the attack is on request headers (not responses). And modern browsers disable
compression on requests (or chrome, maybe even smarter tricks in newer
versions). In fact, Chrome 21 and Firefox 15 switched off SPDY header
compression[1] and the requirements page on discourse:

	https://github.com/discourse/discourse#requirements

shows newer browsers than those are needed. So it seems safe to enable spdy
header compression on the server.

[1]: https://www.imperialviolet.org/2012/09/21/crime.html