FIX: Quote all strings in discourse.conf
Fixes the problem of DB passwords with hashes in them (including passwords
that start with a hash, as reported in https://meta.discourse.org/t/76409).

The problem of passwords with single quotes in them still exists, and will
require modifications to Discourse's GlobalSetting::FileProvider parser,
because the current format doesn't support escaping quote characters.  That
work will have to be deferred to a later time.