FIX: remove pull for Discourse core
We are already shallow-fetching a few lines below. A pull with a shallow
clone can be dangerous if Discourse is using a different repository
or version, as that potentially results in more data being pulled or a
dirty merge with a different upstream.

Remove the pull and rely only on the fetch and checkout.