DEV: add pg and redis gem for protocol-aware healthcheck
The cache_critical_dns script is going to be updated to perform a
healthcheck on the Redis and Postgres hosts prior to caching the
hostname and IP address resolved via DNS. In order to perform a
protocol-aware healthcheck (i.e. a PG/Redis PING-PONG transaction) the
pg and redis gems are required. Ensure the gems are only installed if
the DNS caching behaviour is enabled via templates.