DEV: Build base images for ARM64 platform in a seperate daily job
Why this change?

We're currently building a base image for arm64 platforms as part of the
CI workflow. However, building it on Github action via emulation is too
slow and makes it meaningless to include as part of the CI workflow.

What does this change do?

This change moves the building of the bage image targeting the arm64
platform completely to a scheduled job that runs outside of the normal
CI workflow. This will also enable us to build the images for the test
and dev environment eventually as part of the workflow which we intend
to do in a follow up commit.