DEV: Update standalone.yml
Improve SMTP guidance in the config file

Updated the template advising that the sending domain or sub-domain and DISCOURSE_NOTIFICATION_EMAIL address should be authenticated with SMTP providers to ensure email delivery. 

Also provided a direct link to info on the `discourse-doctor` command that can assist with testing and debugging email configurations.