discourse-doctor: support custom app names as CLI arg
While `launcher` allows to pass any container/app name as CLI argument, `discourse-doctor` does not, but support two hardcoded names only.

This commit adds support for custom names taken from first CLI argument. It is done in a non-breaking change. If the argument is empty, or the related YAML file does not exist, it continues to check for the two hardcoded names like before.

Signed-off-by: MichaIng <micha@dietpi.com>