DEV: Make logrotate conf compatible with Pitchfork
As we’re migrating from Unicorn to Pitchfork, we need to do some
adjustments. Indeed, Pitchfork doesn’t handle logs at all and delegates
this to the underlying system. Therefore, it doesn’t respond to signals
like Unicorn when rotating its logs.

The solution is to use `copytruncate` which works with any app.